so i do need to do the cutoff now to avoid artifacts.
Matching sound chip recordings by ear is apparently not something I am good at, and neither is reading documentation in its proper context. It's easy to lose that when you have ctrl-f on your side. 

This is probably less wrong than the previous envelope attempts but remember just because it sounds better doesn't mean it is better, and you may not remember the reason for your tweaks in 4 years if you're not still lookin at the original documentation.